requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
PyGithub
openai